update gadgets dependency, basic local storage support working.
short explanation of what this all does:

fiatjaf, [11/4/25 10:58 AM]
download a bunch of events from people we follow at startup ("sync"), store them in a local indexeddb that I made for gadgets that is pretty fast, keep track of timestamps of the newest and oldest notes for each pubkey, such that when paginating back we know for which keys we have to ask relays for more notes (using the "until" parameter), and keep a live subscription for all these people that saves to the local db and also updates the timestamps

fiatjaf, [11/4/25 10:59 AM]
do the same for profiles we visit but that aren't followed, but this time do the syncing phase at runtime

fiatjaf, [11/4/25 11:00 AM]
the syncing part is done in a slow pace to not get denied by relays, but doesn't matter since it's very fast anyway except on the first run (in this case I'm tempted to do a preliminary fetch just to show something, and then start the slow sync)

fiatjaf, [11/4/25 11:04 AM]
eventually we can support using negentropy for the syncing phase when possible

fiatjaf, [11/4/25 11:05 AM]
all the other feeds are done with normal relay queries as they are (by the way, there is a disgusting amount if queries made only to hardcoded "big relays" that slowly we should turn into less centralizing logic)

---

now, really, the core logic change was rewriting `subscribeTimeline()` and `loadMoreTimeline()` to take a mix of "relays" and "local" sourced queries, and handle both, even mixing results if necessary.

plus the background syncing for followed people.

for now only the "following" feed and individual profile feeds are served from "local", everything else is served from "relays".

in the rewrite many things were included: stop using ad-hoc relay query functions in favor of functions from `@nostr/tools`, rewrite of the `fetchEvent()` function to be more understandable and less dependent on big relays, other things I forgot.

in the very near future we can start serving replies, reactions, notifications and so on from the local store too -- but mixed with data from relays (and saving the relay data to local store too, in an ad-hoc way, not the methodic sync we do with profiles). that should yield a nice "local-first" approach.