smarter updates when syncing finishes.
whenever there is a background syncing going on, when it ends, `onNew()` will be called now.
it should have been called before but wasn't, which cause the UI to be stale without the newfound
content until the user refreshed manually.

on NoteList and GroupedNoteList `onNew()` was now adapted to perform better when a huge influx of
new notes suddenly arrive (we use `batchDebounce` to do a single updates even though `onNew` is called
hundreds of times), and the UI is now updated smartly:

- new notes that would be put at the top of the feed are delegated to the `newEvents` array, which
gets merged only when the user clicks (these are also sorted now on GroupedNoteList, which is necessary
as we assume notes are sorted in the groups calculation).
- older notes, notes authored by the logged account (or, in the case of GroupedNoteList, notes from authors
that are at the top but that wouldn't be their top notes so wouldn't change the view) are inserted
immediately.